requirements.txt
Flask==3.0.2
pymongo==4.6.1
python-dotenv==1.0.1
Pillow==10.0.0
pytesseract==0.3.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
flask-cors==4.0.0